+++ lib/archethic/self_repair/sync.ex
@@ -2,20 +2,22 @@ defmodule Archethic.SelfRepair.Sync do
   @moduledoc false
 
   alias Archethic.BeaconChain
-  #  alias Archethic.BeaconChain.Summary, as: BeaconSummary
+  alias Archethic.BeaconChain.Subset.P2PSampling
+  alias Archethic.BeaconChain.SummaryAggregate
 
-  alias Archethic.Crypto
   alias Archethic.DB
 
-  alias Archethic.Election
+  alias Archethic.PubSub
 
   alias Archethic.P2P
-  alias Archethic.P2P.Message.GetBeaconSummaries
-  alias Archethic.P2P.Message.BeaconSummaryList
   alias Archethic.P2P.Node
 
-  alias __MODULE__.BeaconSummaryHandler
-  alias __MODULE__.BeaconSummaryAggregate
+  alias __MODULE__.TransactionHandler
+
+  alias Archethic.TaskSupervisor
+  alias Archethic.TransactionChain
+
+  alias Archethic.Utils
 
   require Logger
 
@@ -102,90 +104,146 @@ defmodule Archethic.SelfRepair.Sync do
 
     start = System.monotonic_time()
 
-    authorized_nodes = P2P.authorized_and_available_nodes()
-
-    #  [0, 1, ...]    Subsets
-    #     / | \ 
-    #    /  |  \
-    # [ ]  [ ]  [ ]  Node election
-    #  |\  /|\  /|
-    #  | \/ | \/ |
-    #  | /\ | /\ |
-    # [ ]  [ ]  [ ] Partition by node
-    #  |    |    |
-    # [ ]  [ ]  [ ] Aggregate addresses 
-    #  |    |    |
-    # [ ]  [ ]  [ ] Fetch summaries 
-    #  |\  /|\  /|
-    #  | \/ | \/ |
-    #  | /\ | /\ |
-    # [D1] [D2] [D3] Partition by date
-    #  |    |    |
-    # [ ]  [ ]  [ ] Aggregate and consolidate summaries 
-    #  \    |    /
-    #   \   |   /
-    #    \  |  /
-    #     \ | /
-    #      [ ]
-
-    BeaconChain.list_subsets()
-    |> Flow.from_enumerable()
-    |> Flow.flat_map(fn subset ->
-      # Foreach subset and date we compute concurrently the node election 
-      last_sync_date
-      |> BeaconChain.next_summary_dates()
-      |> Stream.map(&get_summary_address_by_node(&1, subset, authorized_nodes))
-      |> Enum.flat_map(& &1)
-    end)
-    # We partition by node
-    |> Flow.partition(key: {:elem, 0})
-    |> Flow.reduce(fn -> %{} end, fn {node, summary_address}, acc ->
-      # We aggregate the addresses for a given node
-      Map.update(acc, node, [summary_address], &[summary_address | &1])
-    end)
-    |> Flow.flat_map(fn {node, addresses} ->
-      # For this node we fetch the summaries
-      fetch_summaries(node, addresses)
-    end)
-    # We repartition by summary time to aggregate summaries for a date
-    |> Flow.partition(key: {:key, :summary_time})
-    |> Flow.reduce(fn -> %BeaconSummaryAggregate{} end, fn summary, acc ->
-      %{acc | summary_time: Map.get(acc, :summary_time) || summary.summary_time}
-      |> BeaconSummaryAggregate.add_summary(summary)
-    end)
-    |> Flow.emit(:state)
-    |> Stream.reject(&BeaconSummaryAggregate.empty?/1)
-    |> Enum.sort_by(& &1.summary_time)
-    |> Enum.each(&BeaconSummaryHandler.process_summary_aggregate(&1, patch))
+    last_sync_date
+    |> BeaconChain.next_summary_dates()
+    |> BeaconChain.fetch_summary_aggregates()
+    |> Enum.each(&process_summary_aggregate(&1, patch))
 
     :telemetry.execute([:archethic, :self_repair], %{duration: System.monotonic_time() - start})
   end
 
-  defp get_summary_address_by_node(date, subset, authorized_nodes) do
-    filter_nodes =
-      Enum.filter(authorized_nodes, &(DateTime.compare(&1.authorization_date, date) == :lt))
+  @doc """
+  Process beacon summary to synchronize the transactions involving.
 
-    summary_address = Crypto.derive_beacon_chain_address(subset, date, true)
+  Each transactions from the beacon summary will be analyzed to determine
+  if the node is a storage node for this transaction. If so, it will download the
+  transaction from the closest storage nodes and replicate it locally.
 
-    subset
-    |> Election.beacon_storage_nodes(date, filter_nodes)
-    |> Enum.map(fn node ->
-      {node, summary_address}
+  The P2P view will also be updated if some node information are inside the beacon chain to determine
+  the readiness or the availability of a node.
+
+  Also, the  number of transaction received during the beacon summary interval will be stored.
+  """
+  @spec process_summary_aggregate(SummaryAggregate.t(), binary()) :: :ok
+  def process_summary_aggregate(
+        %SummaryAggregate{
+          summary_time: summary_time,
+          transaction_summaries: transaction_summaries,
+          p2p_availabilities: p2p_availabilities
+        },
+        node_patch
+      ) do
+    transaction_summaries
+    |> Enum.reject(&TransactionChain.transaction_exists?(&1.address))
+    |> Enum.filter(&TransactionHandler.download_transaction?/1)
+    |> synchronize_transactions(node_patch)
+
+    p2p_availabilities
+    |> Enum.reduce(%{}, fn {subset,
+                            %{
+                              node_availabilities: node_availabilities,
+                              node_average_availabilities: node_average_availabilities,
+                              end_of_node_synchronizations: end_of_node_synchronizations
+                            }},
+                           acc ->
+      sync_node(end_of_node_synchronizations)
+
+      reduce_p2p_availabilities(
+        subset,
+        summary_time,
+        node_availabilities,
+        node_average_availabilities,
+        acc
+      )
+    end)
+    |> Enum.each(&update_availabilities/1)
+
+    update_statistics(summary_time, length(transaction_summaries))
+  end
+
+  defp synchronize_transactions([], _node_patch), do: :ok
+
+  defp synchronize_transactions(transaction_summaries, node_patch) do
+    Logger.info("Need to synchronize #{Enum.count(transaction_summaries)} transactions")
+    Logger.debug("Transaction to sync #{inspect(transaction_summaries)}")
+
+    Task.Supervisor.async_stream_nolink(
+      TaskSupervisor,
+      transaction_summaries,
+      &TransactionHandler.download_transaction(&1, node_patch),
+      on_timeout: :kill_task
+    )
+    |> Stream.filter(&match?({:ok, _}, &1))
+    |> Stream.each(fn {:ok, tx} ->
+      TransactionHandler.process_transaction(tx)
+    end)
+    |> Stream.run()
+  end
+
+  defp sync_node(end_of_node_synchronizations) do
+    end_of_node_synchronizations
+    |> Enum.each(fn public_key -> P2P.set_node_globally_synced(public_key) end)
+  end
+
+  defp reduce_p2p_availabilities(
+         subset,
+         time,
+         node_availabilities,
+         node_average_availabilities,
+         acc
+       ) do
+    node_list = Enum.filter(P2P.list_nodes(), &(DateTime.diff(&1.enrollment_date, time) <= 0))
+
+    subset_node_list = P2PSampling.list_nodes_to_sample(subset, node_list)
+
+    node_availabilities
+    |> Utils.bitstring_to_integer_list()
+    |> Enum.with_index()
+    |> Enum.reduce(acc, fn {available_bit, index}, acc ->
+      node = Enum.at(subset_node_list, index)
+      avg_availability = Enum.at(node_average_availabilities, index)
+
+      if available_bit == 1 and Node.synced?(node) do
+        Map.put(acc, node, %{available?: true, average_availability: avg_availability})
+      else
+        Map.put(acc, node, %{available?: false, average_availability: avg_availability})
+      end
     end)
   end
 
-  defp fetch_summaries(node, addresses) do
-    addresses
-    |> Stream.chunk_every(10)
-    |> Stream.flat_map(fn addresses ->
-      case P2P.send_message(node, %GetBeaconSummaries{addresses: addresses}) do
-        {:ok, %BeaconSummaryList{summaries: summaries}} ->
-          summaries
+  defp update_availabilities(
+         {%Node{first_public_key: node_key},
+          %{available?: available?, average_availability: avg_availability}}
+       ) do
+    DB.register_p2p_summary(node_key, DateTime.utc_now(), available?, avg_availability)
 
-        _ ->
-          []
-      end
-    end)
-    |> Enum.to_list()
+    if available? do
+      P2P.set_node_globally_available(node_key)
+    else
+      P2P.set_node_globally_unavailable(node_key)
+      P2P.set_node_globally_unsynced(node_key)
+    end
+
+    P2P.set_node_average_availability(node_key, avg_availability)
+  end
+
+  defp update_statistics(_date, 0), do: :ok
+
+  defp update_statistics(date, nb_transactions) do
+    previous_summary_time =
+      date
+      |> Utils.truncate_datetime()
+      |> BeaconChain.previous_summary_time()
+
+    nb_seconds = abs(DateTime.diff(previous_summary_time, date))
+    tps = nb_transactions / nb_seconds
+
+    DB.register_tps(date, tps, nb_transactions)
+
+    Logger.info(
+      "TPS #{tps} on #{Utils.time_to_string(date)} with #{nb_transactions} transactions"
+    )
+
+    PubSub.notify_new_tps(tps, nb_transactions)
   end
 end
